refactor(daily-task): gate by map name, share player metric helpers
