+++ src/vscripts/modules/helper/player-helper.ts
@@ -80,4 +80,30 @@ export class PlayerHelper {
     });
     return count;
   }
+
+  static GetDamageTaken(playerId: PlayerID): number {
+    let damageTaken = 0;
+    for (let victimID = 0; victimID < DOTA_MAX_TEAM_PLAYERS; victimID++) {
+      if (
+        PlayerResource.IsValidPlayerID(victimID) &&
+        PlayerResource.IsValidPlayer(victimID) &&
+        PlayerResource.GetSelectedHeroEntity(victimID)
+      ) {
+        if (PlayerResource.GetTeam(victimID) !== PlayerResource.GetTeam(playerId)) {
+          damageTaken += PlayerResource.GetDamageDoneToHero(victimID, playerId);
+        }
+      }
+    }
+    return damageTaken;
+  }
+
+  // 扣除从虚拟金币库转回的金额，与 end_screen_2.js money 列口径一致
+  static GetTotalGoldEarned(playerId: PlayerID): number {
+    const virtualGoldData = CustomNetTables.GetTableValue(
+      'player_virtual_gold',
+      playerId.toString(),
+    );
+    const transferredBackTotal = virtualGoldData?.transferred_back_total ?? 0;
+    return Math.max(0, PlayerResource.GetTotalEarnedGold(playerId) - transferredBackTotal);
+  }
 }
